+++ jest.config.js
@@ -1,11 +1,7 @@
 module.exports = {
     testEnvironment: "node",
     coverageProvider: "v8",
-    roots: [
-        "./build/tests/",
-    ],
-    testMatch: [
-        "**/?(*.)+(spec|test).+(ts|tsx|js)",
-    ],
+    roots: [ "./build/tests/", ],
+    testMatch: [ "**/?(*.)+(spec|test).+(ts|tsx|js)", ],
     clearMocks: true,
 };