+++ src/core/brokers/MidaBrokerAccount.ts
@@ -1,14 +1,13 @@
 import { MidaBroker } from "#brokers/MidaBroker";
 import { MidaBrokerAccountParameters } from "#brokers/MidaBrokerAccountParameters";
 import { MidaBrokerAccountType } from "#brokers/MidaBrokerAccountType";
-import { MidaBrokerErrorType } from "#brokers/MidaBrokerErrorType";
 import { MidaEvent } from "#events/MidaEvent";
 import { MidaEventListener } from "#events/MidaEventListener";
 import { MidaBrokerOrder } from "#orders/MidaBrokerOrder";
 import { MidaBrokerOrderDirectives } from "#orders/MidaBrokerOrderDirectives";
 import { MidaBrokerOrderStatusType } from "#orders/MidaBrokerOrderStatusType";
 import { MidaSymbolPeriod } from "#periods/MidaSymbolPeriod";
-import { MidaSymbolQuotationPriceType } from "#quotations/MidaSymbolQuotationPriceType";
+import { MidaSymbolPriceType } from "#symbols/MidaSymbolPriceType";
 import { MidaSymbol } from "#symbols/MidaSymbol";
 import { MidaSymbolTick } from "#ticks/MidaSymbolTick";
 import { MidaEmitter } from "#utilities/emitters/MidaEmitter";
@@ -16,45 +15,51 @@ import { GenericObject } from "#utilities/GenericObject";
 
 /** Represents a broker account. */
 export abstract class MidaBrokerAccount {
-    private readonly _id: string;
-    private readonly _ownerName: string;
-    private readonly _type: MidaBrokerAccountType;
-    private readonly _currency: string;
-    private readonly _broker: MidaBroker;
-    private readonly _emitter: MidaEmitter;
+    readonly #id: string;
+    readonly #ownerName: string;
+    readonly #type: MidaBrokerAccountType;
+    readonly #currency: string;
+    readonly #broker: MidaBroker;
+    readonly #emitter: MidaEmitter;
 
-    protected constructor ({ id, ownerName, type, currency, broker, }: MidaBrokerAccountParameters) {
-        this._id = id;
-        this._ownerName = ownerName;
-        this._type = type;
-        this._currency = currency;
-        this._broker = broker;
-        this._emitter = new MidaEmitter();
+    protected constructor ({
+        id,
+        ownerName,
+        type,
+        currency,
+        broker,
+    }: MidaBrokerAccountParameters) {
+        this.#id = id;
+        this.#ownerName = ownerName;
+        this.#type = type;
+        this.#currency = currency;
+        this.#broker = broker;
+        this.#emitter = new MidaEmitter();
     }
 
     /** The account id. */
     public get id (): string {
-        return this._id;
+        return this.#id;
     }
 
     /** The account owner name. */
     public get ownerName (): string {
-        return this._ownerName;
+        return this.#ownerName;
     }
 
     /** The account type (demo or real). */
     public get type (): MidaBrokerAccountType {
-        return this._type;
+        return this.#type;
     }
 
     /** The account currency (ISO code). */
     public get currency (): string {
-        return this._currency;
+        return this.#currency;
     }
 
     /** The account broker. */
     public get broker (): MidaBroker {
-        return this._broker;
+        return this.#broker;
     }
 
     /** Used to get the account balance. */
@@ -71,33 +76,33 @@ export abstract class MidaBrokerAccount {
 
     /**
      * Used to get an order.
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract getOrder (ticket: number): Promise<MidaBrokerOrder | undefined>;
+    public abstract getOrder (id: string): Promise<MidaBrokerOrder | undefined>;
 
     /**
      * Used to get the gross profit of an order (the order must be in open or closed state).
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract getOrderGrossProfit (ticket: number): Promise<number>;
+    public abstract getOrderGrossProfit (id: string): Promise<number>;
 
     /**
      * Used to get the net profit of an order (the order must be in open or closed state).
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract getOrderNetProfit (ticket: number): Promise<number>;
+    public abstract getOrderNetProfit (id: string): Promise<number>;
 
     /**
      * Used to get the swaps of an order (the order must be in open or closed state).
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract getOrderSwaps (ticket: number): Promise<number>;
+    public abstract getOrderSwaps (id: string): Promise<number>;
 
     /**
      * Used to get the commission of an order.
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract getOrderCommission (ticket: number): Promise<number>;
+    public abstract getOrderCommission (id: string): Promise<number>;
 
     /**
      * Used to place an order.
@@ -107,53 +112,53 @@ export abstract class MidaBrokerAccount {
 
     /**
      * Used to cancel an order (the order must be in pending state).
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract cancelOrder (ticket: number): Promise<void>;
+    public abstract cancelOrder (id: string): Promise<void>;
 
     /**
      * Used to close an order (the order must be in open state).
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract closeOrder (ticket: number): Promise<void>;
+    public abstract closeOrder (id: string): Promise<void>;
 
     /**
      * Used to get the stop loss of an order.
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract getOrderStopLoss (ticket: number): Promise<number | undefined>;
+    public abstract getOrderStopLoss (id: string): Promise<number | undefined>;
 
     /**
      * Used to set the stop loss of an order.
-     * @param ticket The order ticket.
+     * @param id The order id.
      * @param stopLoss The stop loss.
      */
-    public abstract setOrderStopLoss (ticket: number, stopLoss: number): Promise<void>;
+    public abstract setOrderStopLoss (id: string, stopLoss: number): Promise<void>;
 
     /**
      * Used to clear the stop loss of an order.
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract clearOrderStopLoss (ticket: number): Promise<void>;
+    public abstract clearOrderStopLoss (id: string): Promise<void>;
 
     /**
      * Used to get the take profit of an order.
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract getOrderTakeProfit (ticket: number): Promise<number | undefined>;
+    public abstract getOrderTakeProfit (id: string): Promise<number | undefined>;
 
     /**
      * Used to set the take profit of an order.
-     * @param ticket The order ticket.
+     * @param id The order id.
      * @param takeProfit The take profit.
      */
-    public abstract setOrderTakeProfit (ticket: number, takeProfit: number): Promise<void>;
+    public abstract setOrderTakeProfit (id: string, takeProfit: number): Promise<void>;
 
     /**
      * Used to clear the take profit of an order.
-     * @param ticket The order ticket.
+     * @param id The order id.
      */
-    public abstract clearOrderTakeProfit (ticket: number): Promise<void>;
+    public abstract clearOrderTakeProfit (id: string): Promise<void>;
 
     /** Used to get the account symbols. */
     public abstract getSymbols (): Promise<string[]>;
@@ -176,7 +181,7 @@ export abstract class MidaBrokerAccount {
      * @param timeframe The periods timeframe.
      * @param priceType The periods price type.
      */
-    public abstract getSymbolPeriods (symbol: string, timeframe: number, priceType?: MidaSymbolQuotationPriceType): Promise<MidaSymbolPeriod[]>;
+    public abstract getSymbolPeriods (symbol: string, timeframe: number, priceType?: MidaSymbolPriceType): Promise<MidaSymbolPeriod[]>;
 
     /**
      * Used to get the latest symbol tick.
@@ -196,9 +201,17 @@ export abstract class MidaBrokerAccount {
      */
     public abstract getSymbolAsk (symbol: string): Promise<number>;
 
-    // public abstract watchSymbol (symbol: string): Promise<void>;
-    // public abstract getWatchedSymbols (): Promise<string[]>;
-    // public abstract unwatchSymbol (symbol: string): Promise<void>;
+    /**
+     * Used to watch the ticks of a symbol.
+     * Do not use this method directly, use a market watcher instead.
+     * @param symbol The string representation of the symbol.
+     */
+    public abstract watchSymbolTicks (symbol: string): Promise<void>;
+
+    /**
+     * Used to disconnect the account.
+     */
+    public abstract logout (): Promise<void>;
 
     /** Used to get the account free margin. */
     public async getFreeMargin (): Promise<number> {
@@ -208,9 +221,7 @@ export abstract class MidaBrokerAccount {
         return equity - usedMargin;
     }
 
-    /**
-     * Used to get the account margin level.
-     */
+    /** Used to get the account margin level. */
     public async getMarginLevel (): Promise<number> {
         const tasks: Promise<number>[] = [ this.getEquity(), this.getUsedMargin(), ];
         const [ equity, usedMargin, ]: number[] = await Promise.all(tasks);
@@ -244,6 +255,7 @@ export abstract class MidaBrokerAccount {
         return this.getOrdersByStatus(MidaBrokerOrderStatusType.CLOSED);
     }
 
+    /* To implement later.
     public async getPlaceOrderObstacles (directives: MidaBrokerOrderDirectives): Promise<MidaBrokerErrorType[]> {
         const obstacles: MidaBrokerErrorType[] = [];
         const symbol: MidaSymbol | undefined = await this.getSymbol(directives.symbol);
@@ -279,6 +291,7 @@ export abstract class MidaBrokerAccount {
 
         return obstacles.length === 0;
     }
+    */
 
     public async tryPlaceOrder (directives: MidaBrokerOrderDirectives): Promise<MidaBrokerOrder | undefined> {
         try {
@@ -289,15 +302,21 @@ export abstract class MidaBrokerAccount {
         }
     }
 
+    public on (type: string): Promise<MidaEvent>
+    public on (type: string, listener: MidaEventListener): string
     public on (type: string, listener?: MidaEventListener): Promise<MidaEvent> | string {
-        return this._emitter.on(type, listener);
+        if (!listener) {
+            return this.#emitter.on(type);
+        }
+
+        return this.#emitter.on(type, listener);
     }
 
     public removeEventListener (uuid: string): void {
-        this._emitter.removeEventListener(uuid);
+        this.#emitter.removeEventListener(uuid);
     }
 
     protected notifyListeners (type: string, descriptor?: GenericObject): void {
-        this._emitter.notifyListeners(type, descriptor);
+        this.#emitter.notifyListeners(type, descriptor);
     }
 }