+++ src/core/quotations/MidaSymbolQuotationParameters.ts
@@ -4,5 +4,4 @@ export type MidaSymbolQuotationParameters = {
     date: Date;
     bid: number;
     ask: number;
-    exchangeName?: string;
 };