+++ src/core/symbols/MidaSymbolPriceType.ts
@@ -1,0 +1,5 @@
+/** The price type of a symbol. */
+export enum MidaSymbolPriceType {
+    BID = "bid",
+    ASK = "ask",
+}