+++ src/core/utilities/cloneable/IMidaCloneable.ts
@@ -1,4 +1,6 @@
+import { GenericObject } from "#utilities/GenericObject";
+
 /** Represents a cloneable object. */
 export interface IMidaCloneable {
-    clone (): any;
+    clone (): GenericObject;
 }