+++ src/core/utilities/emitters/MidaEmitter.ts
@@ -1,17 +1,15 @@
-import { v1 as uuidV1 }  from "uuid";
-
 import { MidaEvent } from "#events/MidaEvent";
 import { MidaEventListener } from "#events/MidaEventListener";
 import { GenericObject } from "#utilities/GenericObject";
+import { v1 as uuidV1 } from "uuid";
 
 /** Represents an entity capable of emitting events to a its listeners. */
 export class MidaEmitter {
-    private static readonly _ANY_TYPE_KEY: string = "*";
-
-    private readonly _listeners: Map<string, Map<string, MidaEventListener>>;
+    static readonly #ANY_TYPE_KEY: string = "*";
+    readonly #listeners: Map<string, Map<string, MidaEventListener>>;
 
     public constructor () {
-        this._listeners = new Map();
+        this.#listeners = new Map();
     }
 
     public addEventListener (type: string, listener: MidaEventListener): string {
@@ -20,39 +18,41 @@ export class MidaEmitter {
         do {
             uuid = uuidV1();
         }
-        while (this._uuidExists(uuid)); // This software deals with real money. Better to avoid even the most improbable things.
+        while (this.#uuidExists(uuid)); // This software deals with money. Better to avoid even the most improbable things.
 
-        const listeners: Map<string, MidaEventListener> = this._listeners.get(type) || new Map();
+        const listenersOfType: Map<string, MidaEventListener> = this.#listeners.get(type) ?? new Map();
 
-        listeners.set(uuid, listener);
-        this._listeners.set(type, listeners);
+        listenersOfType.set(uuid, listener);
+        this.#listeners.set(type, listenersOfType);
 
         return uuid;
     }
 
     public removeEventListener (uuid: string): void {
-        for (const key of this._listeners.keys()) {
-            const listeners: Map<string, MidaEventListener> | undefined = this._listeners.get(key);
+        for (const type of this.#listeners.keys()) {
+            const listenersOfType: Map<string, MidaEventListener> | undefined = this.#listeners.get(type);
 
-            if (listeners && listeners.has(uuid)) {
-                listeners.delete(uuid);
+            if (listenersOfType?.has(uuid)) {
+                listenersOfType.delete(uuid);
 
                 break;
             }
         }
     }
 
+    public on (type: string): Promise<MidaEvent>
+    public on (type: string, listener: MidaEventListener): string
     public on (type: string, listener?: MidaEventListener): Promise<MidaEvent> | string {
-        if (listener) {
-            return this.addEventListener(type, listener);
+        if (!listener) {
+            return new Promise((resolve: any): void => {
+                const uuid: string = this.addEventListener(type, async (event: MidaEvent): Promise<void> => {
+                    this.removeEventListener(uuid);
+                    resolve(event);
+                });
+            });
         }
 
-        return new Promise((resolve: any): void => {
-            const uuid: string = this.addEventListener(type, (event: MidaEvent): void => {
-                this.removeEventListener(uuid);
-                resolve(event);
-            });
-        });
+        return this.addEventListener(type, listener);
     }
 
     public notifyListeners (type: string, descriptor?: GenericObject): void {
@@ -63,26 +63,26 @@ export class MidaEmitter {
             descriptor,
         });
 
-        if (type !== MidaEmitter._ANY_TYPE_KEY) {
-            const listenersOfAny: Map<string, MidaEventListener> = this._listeners.get(MidaEmitter._ANY_TYPE_KEY) || new Map();
+        if (type !== MidaEmitter.#ANY_TYPE_KEY) {
+            const listenersOfAny: Map<string, MidaEventListener> = this.#listeners.get(MidaEmitter.#ANY_TYPE_KEY) ?? new Map();
 
             for (const listener of listenersOfAny.values()) {
                 listener(event);
             }
         }
 
-        const listenersOfType: Map<string, MidaEventListener> = this._listeners.get(type) || new Map();
+        const listenersOfType: Map<string, MidaEventListener> = this.#listeners.get(type) ?? new Map();
 
         for (const listener of listenersOfType.values()) {
             listener(event);
         }
     }
 
-    private _uuidExists (uuid: string): boolean {
-        for (const key of this._listeners.keys()) {
-            const listeners: Map<string, MidaEventListener> | undefined = this._listeners.get(key);
+    #uuidExists (uuid: string): boolean {
+        for (const key of this.#listeners.keys()) {
+            const listeners: Map<string, MidaEventListener> | undefined = this.#listeners.get(key);
 
-            if (listeners && listeners.has(uuid)) {
+            if (listeners?.has(uuid)) {
                 return true;
             }
         }