+++ src/core/watcher/MidaMarketWatcherDirectives.ts
@@ -1,0 +1,5 @@
+export type MidaMarketWatcherDirectives = {
+    watchTicks?: boolean;
+    watchPeriods?: boolean;
+    timeframes?: number[];
+};